+++ src/creatures/players/player.hpp
@@ -1211,6 +1211,7 @@ public:
 
 	bool checkAutoLoot(bool isBoss) const;
 	bool checkChainSystem() const;
+	bool checkEmoteSpells() const;
 
 	QuickLootFilter_t getQuickLootFilter() const;